requirements.txt
line-bot-sdk
flask
werkzeug>=0.15.4
python-dotenv
pydub~=0.25.1
openai